### 2020-04-29 #82 @whedon
[PRE REVIEW]: An open source crash course on parameter estimation of computational models using a Bayesian optimization approach
**Submitting author:** @mbarzegary (<a href="http://orcid.org/0000-0002-1456-0610">Mojtaba Barzegari</a>)
**Repository:** <a href="https://github.com/mbarzegary/educational-bayesian" target ="_blank">https://github.com/mbarzegary/educational-bayesian</a>
**Version:** v1.0.0

@mbarzegary if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
