+++ paper.bib
@@ -4,7 +4,7 @@
   year      = {1987},
   author    = {Albert Tarantola},
   isbn      = {0444427651},
-  url       = {https://www.amazon.com/Inverse-Problem-Theory-Parameter-Estimation/dp/0444427651?SubscriptionId=AKIAIOBINVZYXZQZ2U3A&tag=chimbori05-20&linkCode=xm2&camp=2025&creative=165953&creativeASIN=0444427651},
+  doi       = {10.1137/1.9780898717921},
 }
 
 @Book{Mockus1989,
