### 2020-07-13 #82 @matbesancon
Hi, thanks for pinging me but I'm not sure I am the right  person for this review, I don't focus on Bayesian models
