### 2020-08-12 #89 @whedon
[REVIEW]: An open source crash course on parameter estimation of computational models using a Bayesian optimization approach
**Submitting author:** @mbarzegary (<a href="http://orcid.org/0000-0002-1456-0610">Mojtaba Barzegari</a>)
**Repository:** <a href="https://github.com/mbarzegary/educational-bayesian" target ="_blank">https://github.com/mbarzegary/educational-bayesian</a>
**Version:** v2.0
