### 2020-08-25 #89 @magsol
Hey @e-dub and @jcapriot , wanted to check in and see how things were going. Any kind of ETA or obstacles I can help with?
