### 2020-09-15 #89 @magsol
Hi @e-dub and @jcapriot , how are things going? Anything I can help with?
