### 2020-09-18 mbarzegary/educational-bayesian#1 @e-dub
Release version
Am I missing the release version?  According to the review checklist this is needed. See https://github.com/openjournals/jose-reviews/issues/89
