### 2020-09-30 mbarzegary/educational-bayesian#1 @magsol
Yes, we do indeed need a release version! Thanks @e-dub for pointing this out.
