### 2020-11-17 #89 @mbarzegary
@magsol may I start to answer the comments? or I should wait for the second review to come. it's almost one year since I started this process.
