### 2020-11-18 #89 @e-dub
> @mbarzegary Yes, please respond to the comments! I had been planning to follow up on this thread and see if you even wanted the review to go forward since I hadn't heard anything from you responding to existing feedback. If you want to continue, then yes--please respond to the feedback already given, and I will track down another reviewer if @e-dub does not respond ASAP.

Respond??? See [here](https://github.com/mbarzegary/educational-bayesian/issues)
