### 2020-11-18 #89 @magsol
@e-dub I'm so sorry! I'd blame it on "dad brain" (I'm on paternity leave at the moment so everything is super fuzzy) but I should have just scrolled up a bit further; my apologies.

@mbarzegary Please check out the "issues" tab for your submission, under which @e-dub has submitted numerous pieces of feedback you can begin to address.

> @magsol Sorry for the lack of response, I will get this today and work through it. It looks like the original accept link has expired for me though.

@jcapriot Here is the link https://github.com/mbarzegary/educational-bayesian/
