+++ src/README.md
@@ -1,3 +1,0 @@
-# Source code
-
-The complete source code [is maintained in another GitHub repository](https://github.com/mbarzegary/BayesianFEM)
