### 2020-12-08 e5e781b
remove reference to BayesianFEM code
