### 2020-12-08 f60499d
add CONTRIBUTING.md. fixes #5
