+++ CONTRIBUTING.md
@@ -1,0 +1,39 @@
+# Contributing
+
+**IMPORTANT**: this document has been adapted from the excellent Contributing
+guide of the [UofTCoders' `rcourse`](https://github.com/UofTCoders/rcourse/blob/master/CONTRIBUTING.md).
+
+This repository is an open-source crash course,
+aimed to introduce and teach how to construct inverse problems in computational engineering.
+
+## How to Contribute
+
+Anyone can contribute to the course repository via [pull requests][pull-requests].
+
+We use [GitHub flow][github-flow] to manage changes:
+
+1. Create a [fork][fork-explanation] of this repository, and [clone][clone-explanation] it to your local computer.
+2. In your local copy of this repository, create a new [branch][branch-explanation].
+3. Commit your changes to that branch.
+4. Push the edits on that branch to your fork on GitHub.
+5. Submit a pull request to the master repository (`mbarzegary/educational-bayesian`).
+6. If you receive feedback on your pull request,
+make further commits to the new branch on your fork. These will automatically be added to
+your pull request.
+
+You may wish to look at [How to Contribute to an Open Source Project on GitHub][contribute]
+for more detailed instructions. The [GitHub Glossary][glossary] is also a useful resource that explains
+Git-related terminology.
+
+You can also open an issue if you find a bug or have a suggestion regarding the provided content.
+
+## Format
+
+If you want to add a new module, please pay attention that it should be developed as a Jupyter notebook file (with `ipynb` extension). That's the format this course is built upon. You may have a look at the available sample(s) to have some ideas of how to structure your module.
+
+[branch-explanation]: https://help.github.com/articles/about-branches/
+[clone-explanation]: https://help.github.com/articles/cloning-a-repository/
+[contribute]: https://egghead.io/courses/how-to-contribute-to-an-open-source-project-on-github
+[fork-explanation]: https://help.github.com/articles/fork-a-repo/
+[github-flow]: https://guides.github.com/introduction/flow/
+[glossary]: https://help.github.com/articles/github-glossary/
