### 2020-12-09 803abdd
add more references. fixes #12
