+++ paper.md
@@ -24,9 +24,9 @@ bibliography: paper.bib
 
 # Summary
 
-Parameter estimation is an inevitable and crucial aspect of computational modeling projects, especially the ones that deal with ordinary differential equations (ODE) or partial differential equation (PDE) models. Well-known examples in this regard are the models derived from a basic balance or conservation law, such as mass balance or heat transfer problems. For real-world applications, these equations contain some coefficients that cannot be obtained directly from published scientific materials or experimental studies. One of the best  solutions to this challenge is constructing an inverse problem.
+Parameter estimation is an inevitable and crucial aspect of computational modeling projects, especially the ones that deal with ordinary differential equations (ODE) or partial differential equation (PDE) models. Well-known examples in this regard are the models derived from a basic balance or conservation law, such as mass balance or heat transfer problems. For real-world applications, these equations contain some coefficients that cannot be obtained directly from published scientific materials or experimental studies [@Dehghan2001]. One of the best solutions to this challenge is constructing an inverse problem.
 
-According to [@Tarantola1987], inverse modeling is the use of the actual results of some measurements of the observable parameters to infer the actual values of the model parameters. Put differently, what we want to do is estimating parameters that cannot be directly measured for our computational model. This is also called parameter estimation or model calibration. Indeed, we calibrate our model to act similarly to available experimental data, and then this calibrated model can be used to simulate other scenarios that that haven’t been tested yet in the experiments. This is a common process in a lot of modeling problems in science and engineering.
+According to [@Tarantola1987], inverse modeling is the use of the actual results of some measurements of the observable parameters to infer the actual values of the model parameters. Put differently, what we want to do is estimating parameters that cannot be directly measured for our computational model. This is also called parameter estimation or model calibration [@Carson2014]. Indeed, we calibrate our model to act similarly to available experimental data, and then this calibrated model can be used to simulate other scenarios that that haven’t been tested yet in the experiments. This is a common process in a lot of modeling problems in science and engineering.
 
 Take a simple reaction-diffusion equation as an example, in which the change of the concentration of a sample chemical component $C$ is studied over time. By assuming that the correlated chemical reaction is $A + 2B \rightleftharpoons C$, occurring in a diffusible medium (such as a chemical solution), the PDE to describe the mass transfer phenomenon over time can be written as [@Grindrod1996]:
 
@@ -34,7 +34,7 @@ $$ \frac{\partial [C]}{\partial t} = \nabla . \left( D_C \nabla [C] \right) + k_
 
 in which $[X]$ denotes the concentration of the chemical component $X$, the $D_C$ is the diffusion coefficient of C in the medium, and $k_1$ and $k_2$ are the rates of the forward and backward reactions, respectively. To solve this PDE numerically and get quantitative data (which is the goal of most of the scientific computing projects), we need to know the value of $D_C$, $k_1$, and $k_2$, which is usually hard-to-find in the literature.
 
-As mentioned above, one solution is solve the inverse problem, in which we can use optimization techniques to minimize the difference of the model output and experimental data. Bayesian optimization is one of the most efficient approaches in this regard [@Mockus1989]. `HyperOpt` is a Python package that provides easy-to-use interfaces to implement a Bayesian optimization problem, making it a proper choice for both educational and practical purposes. In this educational module, we used this package to teach the principles of an efficient parameter estimation pipeline.
[…]
