+++ paper.bib
@@ -21,3 +21,36 @@
   year      = {1996},
   author    = {Peter Grindrod},
 }
+
+@inproceedings{Bergstra2013,
+author = {Bergstra, J. and Yamins, D. and Cox, D. D.},
+title = {Making a Science of Model Search: Hyperparameter Optimization in Hundreds of Dimensions for Vision Architectures},
+year = {2013},
+publisher = {JMLR.org},
+booktitle = {Proceedings of the 30th International Conference on International Conference on Machine Learning - Volume 28},
+pages = {I–115–I–123},
+numpages = {9},
+}
+
+@incollection{Carson2014,
+title = "5 - Parameter Estimation",
+editor = "Ewart Carson and Claudio Cobelli",
+booktitle = "Modelling Methodology for Physiology and Medicine (Second Edition)",
+publisher = "Elsevier",
+edition = "Second Edition",
+pages = "83 - 110",
+year = "2014",
+doi = "10.1016/B978-0-12-411557-6.00005-7",
+author = "Paolo Magni and Giovanni Sparacino",
+}
+
+@article{Dehghan2001,
+  title={An inverse problem of finding a source parameter in a semilinear parabolic equation},
+  author={Dehghan, Mehdi},
+  journal={Applied Mathematical Modelling},
+  volume={25},
+  number={9},
+  pages={743--754},
+  year={2001},
+  publisher={Elsevier}
+}
