### 2020-12-09 f1c7fe2
add pedagogy and relevant courses. fixes #3, fixes #8, fixes #17
