### 2020-12-09 1f43fbb
add instructional design. fixes #11, fixes #14
