+++ paper.bib
@@ -46,11 +46,21 @@ author = "Paolo Magni and Giovanni Sparacino",
 
 @article{Dehghan2001,
   title={An inverse problem of finding a source parameter in a semilinear parabolic equation},
-  author={Dehghan, Mehdi},
+  author={Mehdi Dehghan},
   journal={Applied Mathematical Modelling},
   volume={25},
   number={9},
   pages={743--754},
   year={2001},
   publisher={Elsevier}
+}
+
+@article{Chen2015,
+  title={The worked example effect, the generation effect, and element interactivity},
+  author={Chen, O. and Kalyuga, S. and Sweller, J.},
+  journal={Journal of Educational Psychology},
+  volume={107},
+  number={3},
+  pages={689–704},
+  year={2015},
 }
