### 2020-12-09 e14acd8
describe connection to notebook, fixes #10
