### 2020-12-09 83d7507
add getting started section, fixes #4
