### 2020-12-09 039c850
add learning objective, fixes #3, fixes #17
