### 2020-12-10 60553af
fix typos and language stuff, fixes #9
