### 2020-12-10 c819fd2
add getting started section to README
