### 2020-12-10 mbarzegary/educational-bayesian#1 @mbarzegary
Sorry, I missed this point, so a release is published after performing the modifications required for addressing the review comments.
