### 2020-12-11 mbarzegary/educational-bayesian#20 @jcapriot
Information about updating parameter ranges
In the notebook, you simply state "So by averaging the values of best simulations, we can refine the search domain and re-run the optimization." Yet the numbers you update the range to are not the averages of the values and look to have been informed based on the distribution of the runs. An odd thing is when you repeat the range update for the last time, you choose a range of values for c1 from between [5, 8], yet all of the c1 values from the lowest cost function before that were all above 8, which seems counter intuitive.

I think more discussion about how to update the ranges is warranted to really drive home the tutorial part of this work, as that is a big part of this method.
