### 2020-12-11 #89 @magsol
@jcapriot No worries; 2020 has been a barnburner for all of us in many different ways... to put it lightly.

Hmm; I'm relatively new as an editor so this is the first time I've seen this problem. From what I can tell, it looks like you need to be 1) logged in (clearly you are), and 2) accept the invite at this link: https://github.com/openjournals/joss-reviews/invitations .

If that still doesn't solve the problem, let me know and I'll start pinging people up the chain of command.
