### 2020-12-15 #89 @magsol
Hi @openjournals/joss-eics:

We're running into a problem where one of the reviewers can't seem to modify the review checklist in this ticket. I see @jcapriot 's account listed as "Awaiting jcapriot's response" under the "Manage Access" page for `jose-reviews` repository, but [you can see from the above post](https://github.com/openjournals/jose-reviews/issues/89#issuecomment-743464450) that he's not having any success accepting the invitation.

I'm out of ideas. Could you provide some assistance?
