### 2020-12-15 #89 @magsol
@kthyng Oh geez, I'm sorry about that. Autocorrect for the loss :P

@openjournals/jose-eics , [please see this message above](https://github.com/openjournals/jose-reviews/issues/89#issuecomment-745495226) (ignore that I accidentally tagged joss-eics).
