### 2021-02-07 #89 @magsol
Hi all  👋

The reviews are going silent again. **Please respond if you want to continue the review process, otherwise we'll have to end review without acceptance.**

@e-dub: Can you look over the responses to your initial feedback provided by @mbarzegary and verify that they satisfy your initial concerns?

@jcapriot: Please check for an invitation to modify the JOSE GitHub repos again, as I canceled and resubmitted your request 12 days ago. Notify me ASAP if you still cannot perform edits; otherwise, please proceed with your review.
