### 2021-02-21 #89 @mbarzegary
dear @labarba
yes, I confirm that I have addressed them all. I reopened all the issues to be closed by @e-dub after confirmation.
