### 2021-02-21 8d63cf0
update requirements.txt, fixes #18
