+++ requirements.txt
@@ -2,12 +2,12 @@ cycler==0.10.0
 decorator==4.4.0
 future==0.17.1
 hyperopt==0.1.2
+jupyter==1.0.0
 kiwisolver==1.1.0
 matplotlib==3.1.0
 networkx==2.3
 numpy==1.16.4
 pandas==0.24.2
-pkg-resources==0.0.0
 pymongo==3.8.0
 pyparsing==2.4.0
 python-dateutil==2.8.0
