### 2021-02-23 #89 @labarba
Hi @e-dub 👋  — the author says they have addressed all the issues you opened in the target repository. Could you give this a second look now, and see if you can check off (some of) the remaining items in your checklist?
