### 2021-02-23 #89 @jcapriot
@labarba I still don't have the ability to edit my boxes on this page, and I've never actually received the link to be able to join the JOSE repo to edit them. It still says that I do not have permissions, and I have tried many times to gain permissions. I have not seen it come through. I should have my email listed on my GitHub account here. I did add comments and errors in a couple issues on the target repo that look to have been addressed.
