### 2021-04-06 #89 @mbarzegary
@magsol technically speaking, no. the second reviewer has left some of the checkboxes unchecked without creating issues in the repository, so I don't know his comments on the items to fix them. I have fixed most of the unchecked items according to the comments of reviewer 1.
