### 2021-04-22 #89 @magsol
Hi @jcapriot 👋 Please see the note from the author below regarding your review checklist:

> @magsol technically speaking, no. the second reviewer has left some of the checkboxes unchecked without creating issues in the repository, so I don't know his comments on the items to fix them. I have fixed most of the unchecked items according to the comments of reviewer 1.

Please respond as soon as you can with whatever help you may need from me or the author, and whatever additional steps you have yet to take.
