### 2021-05-10 #89 @jcapriot
Hello everyone, I have gone through and updated the checkboxes. There is still an unresolved issue related to my review on the target repo, (from Dec 10th: https://github.com/mbarzegary/educational-bayesian/issues/20) that I feel is the only place where this module is significantly lacking.
