### 2021-05-11 mbarzegary/educational-bayesian#20 @mbarzegary
@jcapriot you are totally right. thank you for notifying this. I change the random seed number several times to find a better starting point that helps describing the process in a more clear manner. then, I added more explanation on tuning the search space for each optimization step. I also discussed its various aspects and the things the students should pay attention to.
