### 2021-05-11 7702dbf
clarifying the process of updating search spaces, fixes #20
