### 2021-05-14 #89 @jcapriot
The improvements made to the tutorial are very helpful in understanding the workflow, and they also now show a consistent pattern to the logic for choosing them. They have addressed all of my concerns now.
