### 2021-05-23 #89 @labarba
Hi @e-dub 👋  — We're hoping to get your final checks on this JOSE submission to proceed to publication. Can you pop in the Review issue and give things a new look? Thank you for your contribution!
