### 2021-06-02 mbarzegary/educational-bayesian#21 @e-dub
General comment to publication. 
Here are my general comments here:

I find this document extremely difficult to read and in fact would recommend a full rewrite.  The motivation is lacking, both in regards to someone who may teach, in addition to those who would be interested in taking this crash course.  In my eyes, a crash course should be independent in it self, this is not the case. I would not know where to start to teach the course with the information provided. How would you break this down? Is this a one-day course? a week? a semester? Have you taught this course? And to whom have you taught it?  This would all add to the quality of both the "paper" and the ipython document.
