### 2021-06-03 #89 @labarba
Hello @e-dub : I have seen the issue you opened in the target repository:
https://github.com/mbarzegary/educational-bayesian/issues/21

Given that the comment is a harsh critique of the work, bereft of constructive suggestions, I have taken the time to read the Jupyter notebook from top to bottom (a careful reading took 34 minutes) . Apart from fairly minor typographical and grammatical booboos that could be fixed in one copy editing pass, I found it to be readable and well put together. I am not an expert on optimization or inverse problems, yet I followed the argument and formed a mental idea of how even I could teach with this material. I would assign the introduction to read ahead of lecture, and would prepare myself to go through the worked example in a live-coding format, with students following along. I estimate this would take about 2 hours of a live session, after which students should indeed get a good sense for how these methods work. A more thorough review might lead me to make suggestions or ask clarifications from the author about code improvements (e.g., I might ask why go through lists when handling the data and not go straight to NumPy arrays). Hardly a "complete rewrite" job!

@e-dub, I ask if you are able to enter into a constructive process with this review, or if you find that you don't see a path for that. We are grateful for your efforts and willingness to make a contribution to JOSE. If you'd prefer to bow out, I appreciate you letting me know sooner rather than later.

@mbarzegary — After reading the lesson, my assessment is that it is fairly short, compared with the size of learning module we target in JOSE. That does not make it unpublishable. I would like to suggest augmenting it slightly with peppered  short student exercises, and adding at the end a brief section describing what a student mini-project might look like, for those interested in more lasting retention.
