### 2021-06-03 #89 @e-dub
I stand by my criticism and reject your comment of "bereft of constructive suggestions".  Constructive suggestions were made both in this cycle as well in the first, including the laundry list that is found in the specific comment in question.

If you find that this criticism overly harsh (which I indeed do not) and that it is unnecessarily holding up a work that is otherwise publication ready, feel free to replace me as reviewer.
