### 2021-06-03 #89 @labarba
Thanks for popping back in here to continue the conversation.

Your post on the issue tracker of the target repo has assessments like:

>I find this document extremely difficult to read and in fact would recommend a full rewrite. The motivation is lacking, both in regards to someone who may teach, in addition to those who would be interested in taking this crash course.

Suggesting a full rewrite is quite a harsh. You say the motivation is lacking: what would make it better? (Is this referring to the lesson or to the paper?).

>In my eyes, a crash course should be independent in it self, this is not the case.

How is it not the case? Despite from the technical jargon contained in the las paragraph of the introduction, I found the lesson self-contained. What is missing, in your assessment?

In other words, the author needs your assessments to be followed by requests.
