### 2021-06-04 #89 @mbarzegary
dear @labarba,
I addressed the issues you mentioned. I added a couple of exercises and a mini-project. Moreover, I modified the last paragraph of the introduction section to make it less technical.
