### 2021-06-16 #89 @arokem
At the request of the EiC and editor for this submission, I have reviewed this article. Here is my assessment.

# Review

## General comments.

This paper introduces materials for a "crash course" in Bayesian optimization. The ideas are clearly important and effective materials for teaching this topic should be helpful. The notebook comprising the materials contains useful examples and walks the learner through these examples step by step, in a helpful manner. The inclusion of exercises is beneficial. After some small debugging (see below), the code runs through with no problem. In general, I find some parts harder to read through because terminology is not always defined and sometimes used inconsistently (e.g., "variables"/"parameters").

## Comments on the notebook:

First a technical comment: Code cell 6 contains an un-needed import of the `hyperopt.random` sub-module. This raised an error on my machine because of changes in more recent versions of hyperopt. So, removing this import will make it possible to run this. The same cell also imports `hyperopt.Trials`, which is also unused. So, this line of code could be simplified to `from hyperopt import hp, tpe, fmin`, without affecting subsequent code.

The first section repeats itself a bit, and it takes a little while until the author tells us what the tutorial will be about. In the end of the section, you finally learn that the goal of the tutorial is specifically to teach Bayesian optimization (as the title also suggests). Though they appear in the paper, I think that it would be great to write a few learning objectives at the top of the notebook, that will help learners and instructors calibrate their expectations. You can find some information about writing effective learning objectives [here](https://ii.library.jhu.edu/2016/07/20/writing-effective-learning-objectives/). The first section is also a bit confusing because of the density of new terms that are introduced, some without an explicit definition (e.g., "calibration", "variables", etc.; see general comments above).

It would be good to change function/variable names to use PEP8 (i.e., snake_case, rather than camelCase). For example: "prepare_data_frame" rather than "prepareDataFrame". This will improve readability and teach learners best practices in software, while also teaching them optimization.

### Minor comments:

- In the first section, the phrase: "best available deals" doesn't make sense.

- There is an extraneous double quote after the sentence: "...especially when we are working with heavy computational models!" in section 2 of the notebook ("Bayesian optimization").

- Section 2: "If you interested" should be "If you are interested".

- The `%matplotlib inline` cell magic is not needed (this has been the default behavior of the notebook for quite a while).

- "The stored data can be plotted using Matplotlib library." => "The stored data can be plotted using the Matplotlib library."

- "I said this to assert that this problem is complex enough to be considered like a case with four unknown parameters." is an awkward turn of phrase. Consider changing to something like: "This demonstrates that the problem we will analyze here is complex enough -- at least as complex as other cases with four unkown parameters" or something like that.

- Code cell 7 erroneously defines the inputs of the rmse functions as lists. This
function would error out if provided with lists as inputs, because `-` is not defined as a binary operator for pairs of lists.

- Code cell 8 uses the built-in function name `input` as a variable name, which should be avoided.

- In the text after cell 10, "`runOptimization` first reset the global `iterNumber` variable" should be "`runOptimization` first resets the global `iterNumber` variable".

- In the text after code cell 13, "real-world axes titles" should be "real-world axes labels".

- In the "Exercises" section, "more extensive objective function" should probably be "more expensive objective function".

- In the "Mini project" section contains the sentence "The, you should be able to run the code and optimize our fitting test-case using Python". I think that first word is not needed?

- Same section, the sentence "If you think this first step is a bit difficult, you have a look at this repository, in which the same code is implemented ready-to-use for any problem." Should be "If you think this first step is a bit difficult, you can have a look at this repository, in which the same code is implemented ready-to-use for any problem."

- Same section: "whether you can obtain the most optimal value of the parameters or not". What does it mean when you can't find optimal parameters?

## Comments on the manuscript:

It would be good to define the prerequisites for the course, in terms of what learners need to know before they learn this material. For example, is knowledge of PDEs required. One might get that impression from reading the example (maybe consider simplifying this example?), but presumably the answer is no.

### Minor comments:

The name of the conference for the Bergstra et al. reference contains the phrase "International Conference on" twice.

Chen et al. is missing a DOI (I believe this should be https://doi.org/10.1037/edu0000018).
