### 2021-06-17 9d73bc8
change variable/function names to PEP8
