### 2021-06-17 f8888c4
fix python code issues
