### 2021-06-17 31947bf
improve introduction section and add learning objective
