### 2021-06-17 212acd7
fix references issues and add missing dois
