+++ paper.bib
@@ -27,7 +27,7 @@ author = {Bergstra, J. and Yamins, D. and Cox, D. D.},
 title = {Making a Science of Model Search: Hyperparameter Optimization in Hundreds of Dimensions for Vision Architectures},
 year = {2013},
 publisher = {JMLR.org},
-booktitle = {Proceedings of the 30th International Conference on International Conference on Machine Learning - Volume 28},
+booktitle = {Proceedings of the 30th International Conference on Machine Learning - Volume 28},
 pages = {I–115–I–123},
 numpages = {9},
 }
@@ -52,7 +52,8 @@ author = "Paolo Magni and Giovanni Sparacino",
   number={9},
   pages={743--754},
   year={2001},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/S0307-904X(01)00010-5}
 }
 
 @article{Chen2015,
@@ -63,4 +64,5 @@ author = "Paolo Magni and Giovanni Sparacino",
   number={3},
   pages={689–704},
   year={2015},
+  doi={10.1037/edu0000018}
 }
