### 2021-06-17 #89 @mbarzegary
dear @arokem,
Thank you for your constructive comments. I addressed them through the following distinct commits:

* [change variable/function names to PEP8](https://github.com/mbarzegary/educational-bayesian/commit/9d73bc828219744c953f829d377389e1f85834d9
)
* [fix python code issues](https://github.com/mbarzegary/educational-bayesian/commit/f8888c45b8637f92684c05a5d0e6c02470b1e6fb
)
* [correct ambiguous sentences](https://github.com/mbarzegary/educational-bayesian/commit/330de6838bb4b1436bb37fe242d487eb2bd87872)
* [improve introduction section and add learning objective](https://github.com/mbarzegary/educational-bayesian/commit/31947bf76105c1d6e381a0aa296392a3fcb60435)
* [add prerequisites and update objectives](https://github.com/mbarzegary/educational-bayesian/commit/146137b89339bf961f26ce86e14c11bdca71b78c)
* [fix references issues and add missing dois](https://github.com/mbarzegary/educational-bayesian/commit/212acd73b50595dc99eaa1be60d25e03e6a86675)

Moreover, I released a new version.
