### 2021-06-22 #89 @magsol
I don't see any misspellings or bad citations, but it does seem like one of the citations lacks a DOI: Grindrod 1996. I can't seem to find the article online, so that may be why, but still worth double-checking.

@mbarzegary Please do the following:
- [ ] make a final pass for any typos I may have missed
- [ ] see if the Grindrod 1996 has a DOI, but if not, no worries (just want to make sure my Google skills aren't failing me)
- [ ] create a tagged release and archive
- [ ] report the version number and archive DOI here

Once you've finished these above items we can proceed with acceptance and publication.
