### 2021-06-22 #89 @mbarzegary
Thanks, @magsol
You are right, the citation doesn't have a DOI, and that's why I didn't provide it.

The version number is **v2.0**: [https://github.com/mbarzegary/educational-bayesian/releases/tag/v2.0](https://github.com/mbarzegary/educational-bayesian/releases/tag/v2.0)
The archive DOI is **10.5281/zenodo.5013865**: [https://doi.org/10.5281/zenodo.5013865
](https://doi.org/10.5281/zenodo.5013865)
