### 2021-06-23 07162bd
fix a minor typo in the paper
