### 2021-06-27 #89 @labarba
Congratulations,  @mbarzegary, your JOSE paper is published 🚀
(thank you for your patience as we navigated reduced capacity of volunteers to contribute)

Huge thanks to the Editor: @magsol, and thank you to the Reviewers: @jcapriot, @arokem 🙏
